Removing NLP-Progress covered sections
Maintainer Rationale: 

We ended up with duplicated effort. 

NLP-Progress is maintained by the very popular academic @seb_ruder, with aligned interest in state of the art results across problems. Awesome-NLP is maintained by software engineers with a strong focus on NLP, with aligned interest in figuring out stable working libraries, tools and data sets. 

With this commit, we are streamlining the repo focus significantly. The fewer things we do, the better we can do them.